+++ crates/uzu/tests/moe/moe_experts_perf_test.rs
@@ -101,7 +101,7 @@ fn run_decode_case(
 
     let num_tiles_k = ((d_ff + K_TILE - 1) / K_TILE) as usize;
 
-    let hidden_buf = alloc_buffer::<bf16>(&ctx, sum_k * d_ff);
+    let hidden_buf = alloc_buffer::<f32>(&ctx, sum_k * d_ff);
     let partial_buf = alloc_buffer::<f32>(&ctx, num_tiles_k * sum_k * d_model);
     let output_buf = alloc_buffer::<bf16>(&ctx, sum_k * d_model);
 
@@ -261,7 +261,7 @@ fn run_two_pass_prefill_case(
 
     let num_tiles_k = ((d_ff + K_TILE - 1) / K_TILE) as usize;
 
-    let hidden_buf = alloc_buffer::<bf16>(&ctx, sum_k * d_ff);
+    let hidden_buf = alloc_buffer::<f32>(&ctx, sum_k * d_ff);
     let partial_buf = alloc_buffer::<f32>(&ctx, num_tiles_k * sum_k * d_model);
     let output_buf = alloc_buffer::<bf16>(&ctx, sum_k * d_model);
